_archive_kubot/collectors/requirements.txt
sqlalchemy==2.0.36
asyncpg==0.29.0
python-dotenv==1.0.1
httpx==0.28.1
requests==2.32.3
apscheduler==3.10.4
tenacity==9.1.2
sentry-sdk==2.16.0
beautifulsoup4==4.12.3
certifi>=2024.2.2